Quick Change to Fix Collision Bug:
Was previously getting occurences of where the ball would bounce off
canvas and get stuck out of the window. So, checked code and saw that our
operations to detect collisions with the top of the canvas was actually
checking if the bottom of the ball's position was less than 0 in y pixel
values (needed to be the top of the ball's position compared).