Only/Final Commit - Short Dev. Process
In addition to the finished PongGame, the following additions/changes were
made:

1. Added wallContainer class.
	1a. Added functionality for grouping of blocks at top of screen.
	Used 2D ArrayList to accomplish this. (Object is composed of 2D
	Array of RectF's)

2. Further implementation/adjustments to PongGame class.
	2a. Altered detection collision method to detect/remove blocks
	from the 2D ArrayList.
		Note: The need to dynamically remove elements from the
		data structure swayed me to use the ArrayList.
	2b. Added some new color schemes/cosmetic improvements.
	2c. Added HighScore feature.
	2d. Improved paddle speed/ball dynamics.
	2e. Provided user feedback screens before/after game concludes
	to notify user of outcome/start of game.

Note: Not happy with the design of the game, more particularly the new class and how
it's not very responsible for it's components and associated behavior (2D
ArrayList). Was in a hurry to get this done.

**Only one commit due to myself forgetting to regularly commit, it was a
short dev process. Need to be more consistent/clear with these
commits/commenting in general**

Had fun working through this game!